BREAKING CHANGE: `@site_cache` won't accept unhashable arguments from now on.

I want to use `@site_cache` in a lot more places, but for that I need it to be fast enough.

Currently following inefficiencies exist:
- `frappe.local.site` is accessed many times, each one is a function call and access to ctxvar
- `json.dump` is used to allow unhashable types, it's _quite slow_.
- `datetime` is used for TTL, which is slightly slower than monotonic time.

This PR fixes all three. Results:

bench_utils_bench_site_cache_many_args: 4.03 us +- 0.02 us -> 932 ns +- 9 ns: 4.32x faster
bench_utils_bench_site_cache_no_arg: 4.04 us +- 0.02 us -> 931 ns +- 10 ns: 4.33x faster
bench_utils_bench_site_cache_with_ttl: 4.71 us +- 0.02 us -> 1.11 us +- 0.01 us: 4.24x faster

Geometric mean: 4.30x faster