@ankush 2024-12-17
`@lru_cache` is still 10x faster. Which means it was originally 40x faster :smile: 

bench_utils_bench_site_cache_no_arg: Mean +- std dev: 106 ns +- 1 ns